delete useless folders
