+++ log/tracer.go
@@ -1,5 +1,9 @@
 package main
 
+import (
+	"context"
+)
+
 type Tracer interface {
 	StartTrace(ctx context.Context) error
 	StopTrace()
